tests/wpt/web-platform-tests/html/semantics/interactive-elements/the-popup-element/popup-anchor-element.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<button id=b1>This is an anchor button</button>
<popup id=p1 anchor=b1>This is a popup</popup>
<button id=b2 popup=p1>This button invokes the popup but isn't an anchor</button>

<script>
  test(function() {
    assert_equals(p1.anchor,b1);
  }, "popup anchor IDL property returns the anchor element");
</script>
