LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-popup-element/popup-anchor-element.tentative-expected.txt

This is an anchor button This is a popup This button invokes the popup but isn't an anchor

FAIL popup anchor IDL property returns the anchor element assert_equals: expected (object) Element node <button id="b1">This is an anchor button</button> but got (undefined) undefined
